apps/api/requirements.txt
fastapi==0.116.1
sentry-sdk[fastapi]==2.25.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.36
alembic==1.16.5
asyncpg==0.30.0
psycopg[binary]>=3.1.0
aiosqlite>=0.20.0
redis[hiredis]==5.2.1
pydantic[email]==2.10.4
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
httpx>=0.27.0
snowflake-connector-python==3.17.0; python_version < "3.14"
litellm==1.55.4
